pyproject.toml
[project]
name = "muscleapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "toolz>=1.0.0",
    "typer>=0.19.2",
]
